get a bit closer to working compositing
just to make sure that things are roughly correct, i'm redrawing the entire
overlay on each damage event.  redraws are missing in some places where
damage events don't get triggered (when clicking on an anchor, for
instance), and the flashes from the redraws are quite visible and annoying,
and both the bare client windows and their reparented/framed counterparts
are getting drawn.